Add a CLI

+++ helveg/InteropGraph.cs
@@ -5,7 +5,7 @@ namespace Helveg
 {
     public struct InteropGraph
     {
-        public InteropGraph(Vector2[] positions, int[] weights)
+        public InteropGraph(Vector2[] positions, float[] weights)
         {
             Positions = positions;
             Weights = weights;
@@ -13,12 +13,12 @@ namespace Helveg
 
         public Vector2[] Positions { get; }
 
-        public int[] Weights { get; }
+        public float[] Weights { get; }
 
         public unsafe struct Raw
         {
             public Vector2 *Positions;
-            public int *Weights;
+            public float *Weights;
             public int Count;
         }
     }
